app/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.5
django-debug-toolbar==3.2.2
django-filter==21.1
django-resized==0.3.11
django-rest-auth==0.9.5
djangorestframework==3.12.2
drf-yasg==1.20.0
factory-boy==3.2.1
Faker==9.8.1
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
install==1.3.4
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.4
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.2
Pillow==8.4.0
psycopg2==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
uritemplate==4.1.1
urllib3==1.26.7
